+++ lib/burndown_chart.rb
@@ -14,6 +14,8 @@
 #
 #  To contact SUSE about this file by physical or electronic mail,
 #  you may find current contact information at www.suse.com
+require 'fileutils'
+
 class BurndownChart
 
   attr_accessor :data