+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master (unreleased)
 
 * Find and remove 'Unplanned' labels on `cleanup-sprint`. Fixes #72.
+* Fix `trollolo burndown-init` and `trollolo backup`.
 
 
 ## Version 0.0.11