+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Find and remove 'Unplanned' labels on `cleanup-sprint`. Fixes #72.
 * Fix `trollolo burndown-init` and `trollolo backup`.
+* Remove Ruby 2.1 support. Closes #96.
 
 
 ## Version 0.0.11